refactor: AnalysisLayout 의 불필요한 코드 수정 및 구조 변경
