+++ src/components/layouts/common/TopBar/buttons/FilterButton/FilterButton.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import FilterAltIcon from "@mui/icons-material/FilterAlt";
-import RoundedButton from "../../../../../../components/common/RoundedButton";
-import FilterDrawer from "@containers/home/common/TopBar/buttons/FilterButton/FilterDrawer.tsx";
+import RoundedButton from "../../../../../common/RoundedButton.tsx";
+import FilterDrawer from "@components/layouts/common/TopBar/buttons/FilterButton/FilterDrawer.tsx";
 
 function FilterButton() {
   const [bottomDrawerOpen, setBottomDrawerOpen] = useState(false);
@@ -18,4 +18,5 @@ function FilterButton() {
     </>
   );
 }
+
 export default FilterButton;
