+++ src/components/layouts/common/TopBar/buttons/FilterButton/index.tsx
@@ -1,0 +1,3 @@
+import FilterButton from "./FilterButton.tsx";
+
+export default FilterButton;
