+++ src/components/layouts/common/TopBar/buttons/NotificationButton.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import NotificationsOutlinedIcon from "@mui/icons-material/NotificationsOutlined";
-import RoundedButton from "../../../../../components/common/RoundedButton";
-import { PATH } from "../../../../../constants/path.ts";
+import RoundedButton from "../../../../common/RoundedButton.tsx";
+import { PATH } from "@constants/path.ts";
 
 function NotificationButton() {
   const navigate = useNavigate();
