+++ src/components/layouts/common/TopBar/buttons/PersonalButton.tsx
@@ -2,8 +2,8 @@ import { useNavigate } from "react-router-dom";
 import LoginIcon from "@mui/icons-material/Login";
 // import PersonIcon from '@mui/icons-material/Person';
 import PersonOutlineOutlinedIcon from "@mui/icons-material/PersonOutlineOutlined";
-import RoundedButton from "../../../../../components/common/RoundedButton";
-import { PATH } from "../../../../../constants/path.ts";
+import RoundedButton from "../../../../common/RoundedButton.tsx";
+import { PATH } from "@constants/path.ts";
 import { useSelector } from "react-redux";
 import { selectUser } from "@redux/slices/userSlice.tsx";
