+++ src/components/layouts/common/TopBar/headerMode/HomeMode.tsx
@@ -1,10 +1,10 @@
 import { Stack } from "@mui/material";
 import { useNavigate } from "react-router-dom";
-import LogoButton from "../buttons/LogoButton";
-import PersonalButton from "../buttons/PersonalButton";
+import LogoButton from "../buttons/LogoButton.tsx";
+import PersonalButton from "../buttons/PersonalButton.tsx";
 import FilterButton from "../buttons/FilterButton";
 import SearchButton from "../buttons/SearchButton";
-import NotificationButton from "../buttons/NotificationButton";
+import NotificationButton from "../buttons/NotificationButton.tsx";
 
 function HomeMode() {
   const navigate = useNavigate();
@@ -32,4 +32,5 @@ function HomeMode() {
     </>
   );
 }
+
 export default HomeMode;
