--- src/containers/home/common/TopBar/buttons/FilterButton/index.tsx
@@ -1,3 +1,0 @@
-import FilterButton from "./FilterButton";
-
-export default FilterButton;
